RefactorL Migrate PackActivites to ActivityDirectivesTable
Migrate everything related to the packActivitiesDirectives display and modal to ActivityDirectivesTable. Do based off feedback on bulk shifting activity directives PR
